requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.40
aiosqlite==0.21.0
asyncpg==0.30.0
alembic==1.15.2
httpx
python-dotenv==1.1.0
pydantic-settings==2.9.1
python-jose[cryptography]==3.4.0
livekit>=1.1.3
livekit-agents==1.5.1
livekit-api
google-genai>=1.0.0
openai>=2.0.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
twilio==9.4.4
stripe==11.5.0
posthog==3.8.3
